+++ grace_addon/files/general/www/public/sops.php
@@ -72,7 +72,7 @@
                             <tr>
                                 <td>${file.original_filename}</td>
                                 <td>${file.upload_date}</td>
-                                <td><a href="uploads/sops/${file.unique_filename}" download><i class="fa-solid fa-download"></i> Download</a></td>
+                                <td><a href="download.php?category=sops&file=${encodeURIComponent(file.unique_filename)}" download><i class="fa-solid fa-download"></i> Download</a></td>
                             </tr>
                         `);
                     });